Correct burst value for airguard 52mm turret
A quick check of the OG shows that the twin barrelled airguard only bursts two rounds at a slower rate than it's single barrelled, 40mm, airbourne cousin.

I'll have to fix the fire rate for all weapons once we finally get a fix to that issue on the tracker. But at least this fixes the burst for now.